requirements.txt
flask==3.0.2
speedtest-cli==2.1.3 